requirements.txt
PyQt5
torch
Pillow
mss
numpy
opencv-python
ultralytics
pywin32
keyboard